imp(error): add Display and Error trait to ClapError
