+++ src/lib.rs
@@ -3,16 +3,8 @@
 mod arg;
 mod app;
 mod macros;
+mod error;
 
 pub use arg::{Accumulator, Rule};
 pub use app::{App, AppBuilder, CollectionMatcher};
-
-// ClapError
-#[derive(Debug)]
-pub enum ClapError<'a> {
-    TooManyInstances(&'a str),     // -v #3; -vvvv; 4 > 3
-    ExpectedValue(&'a str),        // --config <config>; --config; missing <conf>; iterator starved
-    UnexpectedLong(String) ,       // unknown long provided
-    UnexpectedShort(char),         // unknown short provided
-    UnexpectedPositional(String),  // unknown positional provided
-}
+pub use error::ClapError;
